Separat as AbstructSingleDecal
