+++ Runtime/Decal/AbstructSingleDecal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bc7bbe7fa9ab25445997ab4c09055d75
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
